decentralization for Admin in "Product" function to delete products

+++ src/Controller/AdminController.php
@@ -133,10 +133,16 @@ class AdminController extends AbstractController
      * @Route("/product/delete/{id}",name="product_delete",requirements={"id"="\d+"})
      */
     
-     public function deletePro(Request $request, Product $p): Response
+     public function deletePro(Request $request, Security $security, Product $p): Response
      {
-         $this->repo->remove($p,true);
-         return $this->redirectToRoute('product_show', [], Response::HTTP_SEE_OTHER);
+        if ($security->isGranted('ROLE_ADMIN')) {
+            $this->repo->remove($p,true);
+            return $this->redirectToRoute('product_show', [], Response::HTTP_SEE_OTHER);
+        }
+        else{
+            return $this->render("error_admin.html.twig",[
+            ]);
+        }
      }
